requirements.txt
flask==3.1.0
celery==5.4.0
python-dotenv==1.0.1
gunicorn==23.0.0
pymongo==4.7.2
requests==2.32.3
